Added custom pilot function

+++ src/com/skilldistillery/jets/JetImpl.java
@@ -3,7 +3,7 @@ package com.skilldistillery.jets;
 public class JetImpl extends Jet {
 
 
-	public JetImpl(String model, int speedInMPH, int rangeInMiles, int price) {
+	public JetImpl(String model, double speedInMPH, double rangeInMiles, int price) {
 		super(model, speedInMPH, rangeInMiles, price);
 	}
